+++ node/directives/validateAdminUserAccess.ts
@@ -41,15 +41,17 @@ export class ValidateAdminUserAccess extends SchemaDirectiveVisitor {
         userAgent,
       }
 
-      const { hasAdminToken, hasValidAdminToken, hasValidAdminTokenFromStore } =
-        await validateAdminToken(context, adminUserAuthToken as string)
+      const { hasAdminToken, hasValidAdminToken } = await validateAdminToken(
+        context,
+        adminUserAuthToken as string,
+        metricFields
+      )
 
       // add admin token metrics
       metricFields = {
         ...metricFields,
         hasAdminToken,
         hasValidAdminToken,
-        hasValidAdminTokenFromStore,
       }
 
       // allow access if has valid admin token
@@ -68,7 +70,7 @@ export class ValidateAdminUserAccess extends SchemaDirectiveVisitor {
 
       // If there's no valid admin token on context, search for it on header
       const { hasAdminTokenOnHeader, hasValidAdminTokenOnHeader } =
-        await validateAdminTokenOnHeader(context)
+        await validateAdminTokenOnHeader(context, metricFields)
 
       // add admin header token metrics
       metricFields = {
@@ -91,15 +93,16 @@ export class ValidateAdminUserAccess extends SchemaDirectiveVisitor {
         return resolve(root, args, context, info)
       }
 
-      const { hasApiToken, hasValidApiToken, hasValidApiTokenFromStore } =
-        await validateApiToken(context)
+      const { hasApiToken, hasValidApiToken } = await validateApiToken(
+        context,
+        metricFields
+      )
 
       // add API token metrics
       metricFields = {
         ...metricFields,
         hasApiToken,
         hasValidApiToken,
-        hasValidApiTokenFromStore,
       }
 
       // allow access if has valid API token