+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add enforcement of new validation for admin and api tokens
+- Add more details to admin and api token validation metric
+
 ## [1.43.1] - 2024-07-24
 
 ### Changed